Use libigl/imgui for visualization in examples
Also noteworthy that there was a bug in ply io for writing meshes where we forgot to include the uchar (number of indices) for each face. The ply mesh writing is not tested, so it's definitely a todo to test it.

In this commit, we don't yet add the algorithms for computing point cloud densities, filtering, noise, etc, but they are implemented in the examples. We'll add them to the library in future commits in the form of common algorithms.

+++ include/pcp/algorithm/surface_nets.hpp
@@ -744,7 +744,7 @@ auto surface_nets(
          grid.y + static_cast<scalar_type>(grid.sy) * grid.dy,
          grid.z + static_cast<scalar_type>(grid.sz) * grid.dz}};
 
-    std::size_t constexpr edges[12][2] = {
+    std::uint8_t constexpr edges[12][2] = {
         {0u, 1u},
         {1u, 2u},
         {2u, 3u},
